requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
aiosqlite==0.19.0
pydantic>=2.7.0
tenacity==8.2.3
python-multipart==0.0.9
httpx==0.24.1 